Small changes to hitbox script, updated stage, fireballs can go left or right.
I made animations for the BoxCollider2Ds of the hitboxes before. I made a fireball hitbox script and a generic hitbox to work with these box colliders. The two scripts are the same, except the generic one gets the data from the animated BoxCollider2D component of the object every frame.

These colliders are fairly offset atm. Will fix this soon.