Imports new music, small update to animation sprites.
Implements the Player "Hit" sprite so that when the player is hit by an attack their sprite changes through use of an animation.
This animation isn't able to be played two times in a row for some reason. I'll investigate that further.
Also makes Nico's left and right attack use the sword sprite.